refract config reading
