+++ src/config/config.yaml
@@ -32,21 +32,21 @@ token:
   contract_addr: "0xc08cd26474722ce93f4d0c34d16201461c10aa8c"
 
 social:
-  # twitter
-  mode: ""
-  username: ""
-  password: ""
-  api_key: ""
-  api_key_secret: ""
-  access_token: ""
-  token_secret: ""
-  monitor_window: 0
-  # discord
-  api_token: ""
-  # telegram
-  bot_token: ""
-  channel_id: 0
-  debug: false
+  twitter:
+    mode: ""
+    username: ""
+    password: ""
+    api_key: ""
+    api_key_secret: ""
+    access_token: ""
+    token_secret: ""
+    monitor_window: 0
+  discord:
+    api_token: ""
+  telegram:
+    bot_token: ""
+    channel_id: 0
+    debug: false
 
 web:
   port: 8000
