+++ src/pkg/clients/twitter_scraper.go
@@ -3,6 +3,7 @@ package clients
 import (
 	"context"
 	"fmt"
+	"github.com/carv-protocol/d.a.t.a/src/internal/conf"
 	"time"
 
 	twitterscraper "github.com/tyxben/twitter-scraper"
@@ -11,12 +12,12 @@ import (
 // TwitterScraper represents a Twitter scraper using browser automation
 type TwitterScraper struct {
 	scraper *twitterscraper.Scraper
-	config  *TwitterConfig
+	config  *conf.TwitterConfig
 	userID  string // Store logged in user's ID
 }
 
 // NewTwitterScraper creates a new Twitter scraper with improved error handling and validation
-func newTwitterScraper(config *TwitterConfig) (*TwitterScraper, error) {
+func newTwitterScraper(config *conf.TwitterConfig) (*TwitterScraper, error) {
 	// Validate config
 	if err := validateConfig(config); err != nil {
 		return nil, fmt.Errorf("invalid twitter config: %w", err)
